Fix broken images when deleting tile or orientation.
The custom renderes for tile and orientation did not take into account
what happens when they're deleted.

For tile, it just needs to check that the value isn't "empty"
(empty string, undefined, or null).

For orientation, since render happens before afterChange, the orientation
value can still be set when the tile value is empty. So the solution is
confirm both the value and the tile ID are not empty.