Add an interactive map for placing tiles.
The map will be a new tab. When you click on a space on the board, it is
translated into the corresponding coordinates. This populates a modal,
which will appear and display all the legal tiles for this space. When you
click on a tile, two things happen: it is highlighted, and all legal
orientations of the tile will appear in the next column. When you select
an orientation, it will be highlighted, and the "Place" button will become
enabled. Clicking it will place the selected tile in the selected
orientation in the appropriate coordinate on the map.

This is just an interactive for visualizations. It is not connected to the
board input mechanism (yet).