Allow deleting a tile on the map.
This change allows deleting from the map in 2 ways. When you click on a
coodinate, there is now a delete button at the top of the modal, which
will remove the tile from the map and the table. Additionally, if you
delete any part of a row in the table, the corresponding tile will be
removed from the map as well.