Process table changes by row.
Whenever a cell changes in the placed tiles table, it may have an impact
on the rest of the row. Namely, it could wipe (or enable) another cell.
As such, if the changes are handled in the wrong order, it could cause them
to be wiped out immediately after being applied.

The easiest way to handle this is to group all the changes by row, and
process each row as a group.