Add support for private companies.
Creates s separate table on the railroads tab for private company entry.
For each supported company, you indicate the owner, which must be an
established railroad, and the coordinate on which its token is placed (if
appropriate). The supported companies are the Seaport Company, Meat
Packing Company, and Mail Contract; the 3 which impact route value.