Add issue reporting from the app.
This takes the form of 3 different buttons.
- top-right corner of the screen, intended for reporting general issues
  with the app
- bottom-left corner of the tile placement modal, intended for reporting
  issues that occur when attempting to place a tile
- bottom-left corner of the route calculation modal, intended for
  reporting issues that occur when calculating route values

All 3 gather relevant information from the state of the app to pass along.
They also require the user to elaborate on the issue, and given them the
chance to enter their email for me to follow up with them. The reports
are then generated and emailed to the address in the BUG_REPORT_EMAIL
environment variable.