Tile orientation list should follow modal viewport.
When scrolling would take the bottom of the tile orientation list above
the bottom of the modal viewport, instead set it to the bottom of the
modal viewport. Same for the top of the tile orientation list, but
reversed.

Also, reset the tile orientation list when new tiles are loaded, such
that the top of the list is at the top of the modal viewport.