Redesigning railroads tab.
The main goal of this redesign is to no longer use Handsontable for the
railroads table. As flexible as it is, it's missing a crucial component
for this table: a multi-select dropdown.

I found a third-party implementation (which wrapped the jQuery choices UI
element), but it was missing some features, so required customization.
Even then, there were a number of bugs around copy/paste and keyboard
entry. Additionally, that component doesn't support choosing the same
option multiple times, rendering it useless for trains. All this resulted
in a messy and suboptimal user experience.

The replacement is a custom table, driven by user interaction and
JavaScript. Users select values using dynamic dropdowns and modals. Future
commits will bring it into feature parity with the previous table, without
the bugs or awkward interaction.