Implement import/export.
Adds a modal with a textarea for entering railroad data in a CSV-like
syntax, delimited by semi-colons. The textarea is pre-populated with any
data present in the railroads table. When saving, the railroad table
contents are replaced with the contents of the import textarea. First,
that data is validated as if it was entered via the UI; any invalid data
is discarded. Then, the variouos fields are set using a set of methods
common to the UI, to ensure consistent state.

This text area acts as the source of the table, although the source is
really the table itself.