Allow calculate as long as one train is valid.
If there are other trains in the table which are not yet ready to be
calcaulated (i.e. they're missing a train, a station, or both), that's
fine. Still allow calculation of the ready railroads. Additioanlly, allow
the railroads which are in the table to be given ownership of private
companies, even if they're not ready.