Introducing import/export feature to map.
This replaces the previous tile placement table (which had its own tab).

The table worked to get this project started, but proved to not be ideal
over time. Various problems arose; pasting was fragile, its API didn't
work as expected, and more. Most of all, keeping it in sync with the map
wasn't straightforward.

This replaces all that with an array and a textarea. The array is used to
track the actual board state, which is in turn used to draw the map. It
also pre-populates the import text area. The textarea contains a CSV-like
representation of the board, with fields separated by semi-colons (;).