Make tile selector keyboard friendly.
TAB to cycle through the tiles, hit ENTER to select and display the
orientations. The focus will also kick over to the first orientation
options. The orientation options are inserting into the TAB sequence after
the tile they're connected to, from either direction. Hit ENTER on an
orientation tile to select and place it.