Add WASD control.
Since this is a good sized hex map, tabbing through all the tiles is a
pain. Adding directional movement eases that. I use the classic WASD
layout, as many people are familiar with that, and there's little
benefit from keeping it on the homerow (a la HJKL) when you're
switching between meatspace and the computer.

I chose for down to move left one tile (instead of right) because that
aligns with the board layout; going down-left from the left-most tile on
the top row, you never encounter a blank space.