Add local storage to private companies table.
The plumbing was designed following the import/export pattern of the
railroads table and the map, even though it isn't adding import/export
functionality. This will make adding it later, if desired, very easy.

Also, note that it loads the railroad names itself, and uses them to
validate the owners. This means it doesn't have to wait for the railroads
table to finish loading.