Display stations on the map.
Loops through the railroads table to gather the coordinates and which
station to draw. The number of tokens at a single coord is tracked to
determine layout. The tile capacity and orientation also dictate how to
layout the tokens. This info is gathered through a new API call, which
queries and compiles info about the tile/space.

All tokens on a tile align with the stations circle; stations on built-in
board spaces are drawn in the center.