Place tokens in the proper location.
When placing a station on a space with no tile, make sure the token is on
the station circle. This is accomplished using a JSON file containing the
offsets (in multiples of token size) in X and Y directions. For Chicago,
there is a sub-object which gives the offset by station neighbor.