Add support for independent railroads.
Updates routes1846 to 0.16.1, which supports independent railroads. Adds a
row to the private companies table for each, as well as their unique logic.

They always display their coordinate, as it's preset and unchangeable.
When an owner is selected, that railroad is given a station in place of
the independent railroad's home city. If the ownership is changed (e.g. to
correct a mistake), the station is removed from the previous owner and
given to the new owner. A railroad who owns an independent railroad cannot
deselect that station in the stations modal.