Condense import/export into a single modal.
Now that all 3 tables are on the same page, it makes sense to unify
their import export UI. The code is the same, but the user only needs to
open up one modal to import (or export) the data for all 3. This is done
by creating a tab in the modal for each import/export text area.