Display city names for private companies.
This applies to both the table and the dropdowns. In both, the city name
is followed by its coordinate.

Due to Chicago Connections as option for the Steamboat Company, the table
had to be significantly widened, which was mostly fit by decreasing the
gap separating it from the railroads table. Additionally, the name "Chicago
Connections" was shortened.