Fix tile selctor scroll issue.
An unknown change (suspected to be in how Chrome handles CSS or a
Bootstrap bugfix) resulted in the placed tile modal having an unbounded
heihgt. This caused the tile orientations to scroll off the screen,
instead of being pinned at the top. The size was re-pinned by setting it
on the modal-content element instead of the modal-body element.