Allow placing station on I5.
I5 (Centralia) is a special space: a built-in phase 4 city. It's one of 5
such spaces. However, Centralia is the only one that isn't a home
station. As such, you'd like to manage its stations via the map, like
with all the regular cities.

The easiest way to do this without introducing more special cases is to
explicitly mark I5 as a legal tile, but ensure its treated as having no
legal track upgrdaes. This led to adding a bit that if there are no
legal tiles for a space, the tiles tab should be hidden. The result is
that I5 allows managing the stations, and is the only space to hide the
tiles tab, and default to the stations tab.