Present railroads and private companies on tabs.
The railroads table can grow quite long with all 7 railroads in play,
especially later in the game when many stations have been placed.
Putting these tables on their own tabs reduces the amount of scrolling
needed. Additionally, as the private companies table will be interacted
with less than the railroads table, so it makes sense to deprioritize it
by defaulting to the railroads table.