Import removed railroads.
THe removed railroads import is combined with the railroads import, both
for user convenience and to mirror the underlying library's CSV format
for easy conversion. However, they're stored in their own tables to keep
later separation easy, if it should be required.

If a railroad is marked as both removed and opened, then after import,
only the opened version remains.