Disable app during initial loading.
Loading the map from local storage can take up to 10 seconds. That is
plenty of time for a user to accidentally (or not) modify the game
state, or load up the import/export modal which will be empty, or
otherwise mess with things before loading can complete. Additionally, as
it looks like the app is doing nothing, you may not know there's data to
be loaded.

During load, we do a couple things. First, we disable all inputs.
Second, we gray out the app, and draw a spinner, to indicate loading.
Upon completion, input functionality is restored and the spinner and
gray out go away.