Simplifying importTiles.
When processing the import tiles text, instead of checking the row's
validity in JavaScript, ask the server for the legal orientations for
the give coordinate and tile ID. If the server responds with a list
of orientations, then check that it includes the given orientation. If
so, the row is valid. If the serer responds with an empty list or null,
or th orientation isn't present, drop it on the floor.

The above procedure cut the time to run importTiles by about 83%.