Draw private company tokens on map.
Specifically, the Steamboat Company and Meat Packing Company. The
location of the token on the tile is determined by a file containing
offsets relative to the token size. They're drawn on the same canvas as
the stations, and are drawn at the same time. Triggers to draw tokens
were added to everywhere the placement could change.