Place private company tokens via map.
There's a 3rd tab which appears on the tile selection modal if a private
company can be placed on the tile. A dropdown for each legal private
company will appear, allowing you to select the owner. If an owner has
been previously set, their name is displayed, next to a delete button,
which allows you to remove their ownership without setting another
railroad's.