Cancel the job when requested.
When the job is started, start a montor process polling for cancellation
requests (via Redis). When a cancellation request comes in, add the
job's ID to the cancelling key in Redis. The monitor request will shut
down this job's worker process, and set the job status appropriately.

The monitor process also checks for job timeouts, and sets the status
appropriately if one occurs.

Cancellations occur when the user either selects another railroad to
calculate, or leaving the calculate tab.