Allow easy migration to 18xx Route Finder.
When a user visits the site, they're greeted with a modal. "Migrate!"
will collect their local storage and send it to the new app, then
redirect them to the 1846 page of the app, with their data loaded. If
they click "Remind Me Later" or "Dismiss", the modal will close. The
difference is "Dismiss" sets a localStorage variable which results in
the modal not opening on the user's next visit.

The user can re-open the modal at any time by clicking the
"Migrate to 18xx" link.